refactor(ts-sdk,vault): move activation-deadline rule into the SDK
