+++ packages/babylon-ts-sdk/src/tbv/core/services/deposit/__tests__/peginState.test.ts
@@ -4,6 +4,7 @@ import {
   canPerformAction,
   ContractStatus,
   getPeginProtocolState,
+  isActivationDeadlinePassedOnChain,
   PeginAction,
 } from "../peginState";
 
@@ -162,6 +163,41 @@ describe("peginProtocolState", () => {
       const state = getPeginProtocolState(ContractStatus.ACTIVE);
       expect(
         canPerformAction(state, PeginAction.SIGN_PAYOUT_TRANSACTIONS),
+      ).toBe(false);
+    });
+  });
+
+  // ==========================================================================
+  // isActivationDeadlinePassedOnChain — mirrors the contract's strict '>'
+  // ==========================================================================
+  describe("isActivationDeadlinePassedOnChain", () => {
+    it("treats the boundary block (current == created + timeout) as NOT expired", () => {
+      expect(
+        isActivationDeadlinePassedOnChain({
+          currentBlock: 1100n,
+          createdAtBlock: 1000n,
+          pegInActivationTimeout: 100n,
+        }),
+      ).toBe(false);
+    });
+
+    it("treats one block past the boundary as expired", () => {
+      expect(
+        isActivationDeadlinePassedOnChain({
+          currentBlock: 1101n,
+          createdAtBlock: 1000n,
+          pegInActivationTimeout: 100n,
+        }),
+      ).toBe(true);
+    });
+
+    it("treats a block below the boundary as NOT expired", () => {
+      expect(
+        isActivationDeadlinePassedOnChain({
+          currentBlock: 1050n,
+          createdAtBlock: 1000n,
+          pegInActivationTimeout: 100n,
+        }),
       ).toBe(false);
     });
   });
